--- src/handlers/auth.py
@@ -1,24 +1,0 @@
-from flask import request
-
-from handlers.handler import Handler
-from handlers import APIResponse
-from handlers import HandlerException
-from spotify import auth
-from session.session import Session
-from session.manager import SessionManager
-
-class AuthHandler(Handler):
-
-    def run(self):
-        code = self.args.get('code');
-        if code is None:
-            raise HandlerException('Missing "code" query parameter!')
-
-        c = self.conn
-        acc, ref = auth.get_tokens(c.client_id, c.client_secret, code)
-        if acc is None or ref is None:
-            raise HandlerException('Failed to retrieve access and refresh tokens')
-
-        id = self.manager.add(Session(acc, ref))
-        
-        return APIResponse(302, {}, f"http://localhost:3000/party/{id}")
