+++ src/web/__init__.py
@@ -1,0 +1,1 @@
+""" web - Web Package """
