+++ src/web/handlers/__init__.py
@@ -14,3 +14,4 @@ class HandlerException(Exception):
     def __init__(self, message):
         Exception.__init__(self)
         self.resp = APIResponse(400, {'message': message}).resp
+
